requirements.txt
boto3==1.36.7
botocore==1.36.7
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
jmespath==1.0.1
joblib==1.4.2
numpy==2.2.2
nvidia-nccl-cu12==2.25.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
urllib3==2.3.0
xgboost==2.1.3
